chore: do not create useless meta for posts / pages in DB [fixes #821]

+++ inc/admin/metabox/controls/control_base.php
@@ -147,18 +147,23 @@ abstract class Control_Base {
 			return;
 		}
 		if ( isset( $_POST[ $this->id ] ) ) {
-			$value = wp_unslash( $_POST[ $this->id ] ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
-			update_post_meta( $post_id, $this->id, $this->sanitize_value( $value ) );
-
-			return;
-		} else {
-			if ( $this->type === 'checkbox' ) {
-				update_post_meta( $post_id, $this->id, $this->sanitize_value( 'off' ) );
+			$value = $this->sanitize_value( wp_unslash( $_POST[ $this->id ] ) ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
+			// Remove post meta on default.
+			if ( $value === $this->settings['default'] ) {
+				delete_post_meta( $post_id, $this->id );
 
 				return;
 			}
+			// Update post meta on other values.
+			update_post_meta( $post_id, $this->id, $value );
+
+			return;
+		} else {
+			// Delete on unset.
+			delete_post_meta( $post_id, $this->id );
+
+			return;
 		}
-		delete_post_meta( $post_id, $this->id );
 	}
 
 	/**
@@ -187,7 +192,7 @@ abstract class Control_Base {
 				return sanitize_text_field( $value );
 				break;
 			case 'range':
-				return sanitize_text_field( $value );
+				return absint( $value );
 				break;
 			case 'input':
 				return sanitize_text_field( $value );
